feat: integrated all parts of the project together
This was completed in just under an hour.
It should not have taken several months for three people who are about to be granted degrees in Computer Science to do this.
This is really upsetting and speaks to the absolute lack of effort displayed by these students.
Word of warning, if you cannot solve problems on your own you will have a very difficult time working in the real world of software engineering.

Good luck to you,
Nick